Update: move ItemSerializer to common module
